base commands (fibonacci)

+++ include/settings.hpp
@@ -2,5 +2,5 @@
 
 #include <cstdint>
 
-using register_type = uint16_t;
-using value_type = uint64_t;
+using register_type = uint32_t;
+using value_type = int64_t;
